Refactor RpcServer tests
